Add files via upload
***This script is meant specifically to run with System authority through RemoteSigned execution policy. 
***This script is designed to work with SYNCROmsp script deployments and their custom variables
This script is primarily to sweep HIPAA devices periodically to ensure Bitlocker drive encryption is turned on. There is some logging built into the script-- mainly for my purposes. It is very useful for policing large numbers of devices that need to be in compliance and storing their bitlocker keys in Syncro Custom Text Fields.
For the reasons above, this script is fairly niche. However, it is broad enough to be repurposed with most any other Remote Management software and some dynamic variables.