M365Tenant_MFA_Check_and_Generate_CSVs.ps1
This will need some tweaking to work, as well as creds for an admin account with access to manage MFA in the organization in question. Please feel free to hit me with any questions. This should take a small amount of tweaking to work for your uses and output report files to your preferred location.

The output are CSV's which can be immediately* reused for bulk-updating MFA in the same organization.