Create Turnoff_PersonalTeams_Win11ChatIcon.PS1
This is an IT issue which some sysadmins working with lower end Windows 11 Pro environments may need to solve. This effectively removes the baked in Windows 11 Personal Version of Teams, as well as the chat icon in the task bar-- this icon will reinstall the personal teams if it's clicked. 

Can be useful if you're deploying Work or School Teams, or want to include this as part of a Windows 11 Debloat Script library.

Hope this saves some sysadmin some time.