M365_MFA_Reset_and_Enable.ps1
*For M365 Security Management* 
Script to take a CSV of users and remove their current authentication method, and set their status back to Enabled, so that they will be asked to enter their MFA information at next sign in.

Example scenarios: 
ex1: You want to remove a business unit's MFA methods because they are all using Phone numbers, and you want to force them to update to MSFT auth

ex2: For some reason there are a  bunch of wrong phone numbers or incorrectly configured auth methods.

Important-- this will not pull users. It requires a CSV ahead of time and sets the contest of it to $users.
There can be variable assignment issues with large CSV's with extraneous columns.

For my purposes I've used a CSV with 4 columns, one of which was UPN and there were roughly 200 records/rows.